double to float line item

+++ db/migrate/20130416135052_add_quantity_to_line_item.rb
@@ -1,5 +1,5 @@
 class AddQuantityToLineItem < ActiveRecord::Migration
   def change
-    add_column :line_items, :quantity, :double
+    add_column :line_items, :quantity, :float
   end
 end
